WLS internal standardization, stage 3
The heavy lifting in the backend is now written. Still need to address
the handshake between the frontend and backend, particularly with
regard to the means.

Currently, means are NA for data of type='acov'. This will need to
change. Old code that does not specify means might need to be broken.
Either we break this old code or we automatically add means of zero to
ordinal variables.  Need to discuss with team and this point.