Fix serious Rampart bugs
Previous code did not consider the rampartScale in determination of
whether the model covariance was identical. This induced small
differences (and large in some models) in evaluation of the fit. In many
cases, the difference was small enough not to arouse suspicion. This bug
only affected datasets with heterogeneous cluster sizes. However, bugs
are seldom alone. Another bug associated with sufficient sets that could
cause memory corruption was fixed as well.